Update README & TODO: remove no redundancy parts
Given the size of the files, I decided to keep things simple and push
everything. It enables to have everything in the same place and see what
the instructions and tests looked like when I added a solution.

It can always change in the future if the repo gets too big.